refactor: 공용 타입 및 DTO를 types 폴더로 정리
