+++ src/components/reservation/modals/PaymentModal.tsx
@@ -4,13 +4,13 @@ import { useEffect, useMemo, useRef, useState } from "react";
 import { useNavigate } from "react-router-dom";
 
 import { requestPayment } from "@/api/endpoints/payments";
-import type { CreateBookingResult } from "@/api/endpoints/reservations";
 import { useConfirmClose } from "@/hooks/common/useConfirmClose";
 import { useModalPresence } from "@/hooks/common/useModalPresence";
 import { useDepositRate } from "@/hooks/reservation/useDepositRate";
 import { useMenus } from "@/hooks/reservation/useMenus";
 import { cn } from "@/lib/utils";
 import { useUserId } from "@/stores/useAuthStore";
+import type { CreateBookingResult } from "@/types/reservation";
 import type { ReservationDraft } from "@/types/restaurant";
 import type { RestaurantDetail } from "@/types/store";
 import { calcMenuTotal } from "@/utils/menu";
