+++ src/types/inquiry.ts
@@ -1,0 +1,4 @@
+export interface ResponseInquiryDTO {
+  id: number;
+  createdAt: string;
+}
